Rename: Pages
- 페이지의 역할을 보다 명확히 표현하기 위해 파일 이름을 변경했습니다.
- Admins to AdminProfile
- Help to CheatModeGuide
- SelectManagers to EditManagers
- InputFoods to InputFoodTemp
- InputMachines to InputMachineTemp
- CustomFoodTemp to RandomFoodTemp
- CustomMachineTemp to RandomMachineTemp
