+++ src/api/index.js
@@ -1,7 +1,8 @@
 import axios from "axios";
 
 const client = axios.create({
-  baseURL: import.meta.env.VITE_SERVER_URL,
+  // 로컬 테스트시 주석
+  // baseURL: import.meta.env.VITE_SERVER_URL,
   timeout: 600 * 1000,
   headers: {
     "Content-Type": "application/json",
