+++ src/pages/InputMachineTemp.jsx
@@ -1,15 +1,15 @@
 import React, { useEffect } from "react";
 import { useLocation } from "react-router-dom";
-import { useCustomFoods, useCustomProducts } from "./../hooks/useProducts";
-import Banner from "./../components/Banner";
-import InputProducts from "./../components/InputProducts";
-import ManagerList from "./../components/ManagerList";
+import { useCustomMachines, useCustomProducts } from "../hooks/useProducts";
+import InputProducts from "../components/InputProducts";
+import Banner from "../components/Banner";
+import ManagerList from "../components/ManagerList";
 import Button from "../components/Button";
-import styles from "./InputFoods.module.css";
+import styles from "./InputTemp.module.css";
 import Modal from "../components/Modal";
 import { useGoHome } from "../hooks/useNavigator";
 
-export default function InputFoods() {
+export default function InputMachineTemp() {
   const location = useLocation();
 
   const { handleClick } = useGoHome();
@@ -17,7 +17,7 @@ export default function InputFoods() {
   const {
     productsQuery: { isLoading, error, data },
     setProductsTemp,
-  } = useCustomFoods();
+  } = useCustomMachines();
 
   const {
     handleSubmit,
@@ -26,30 +26,33 @@ export default function InputFoods() {
     result,
     loading,
     setProducts,
-    setResult,
     selectManager,
     setSelectManager,
+    setResult,
   } = useCustomProducts({ location, setProductsTemp });
 
   useEffect(() => {
-    setProducts(data?.customFood);
+    setProducts(data?.customMachine);
   }, [data]);
 
   return (
     <>
+      {isLoading && <p>Loading...</p>}
       {products && (
         <section className={styles.section}>
           <div className={styles.title}>
-            <div className={styles.text}>식품 입력</div>
+            <div className={styles.text}>기기 입력</div>
             {data?.mgrList && (
               <ManagerList
                 className={styles.mgrList}
                 mgrList={data.mgrList}
+                x
                 selectManager={selectManager}
                 setSelectManager={setSelectManager}
               />
             )}
           </div>
+
           <form
             className={styles.form}
             id='inputMachine'
@@ -84,23 +87,28 @@ export default function InputFoods() {
                 component={"입력에 실패했습니다. 다시 시도해주세요."}
               />
             )}
-            {products.length === 0 ? (
-              <div className={styles.empty}>먼저 식품 선택을 완료해주세요.</div>
-            ) : (
-              products.map((product) => (
-                <div
-                  className={styles.product}
-                  key={product.id}
-                  onSubmit={handleSubmit}
-                >
-                  <InputProducts product={product} />
+            <div>
+              {products.length === 0 ? (
+                <div className={styles.empty}>
+                  먼저 기기 선택을 완료해주세요.
                 </div>
-              ))
-            )}
+              ) : (
+                products.map((product) => (
+                  <div
+                    className={styles.product}
+                    key={product.id}
+                    onSubmit={handleSubmit}
+                  >
+                    <InputProducts product={product} />
+                  </div>
+                ))
+              )}
+            </div>
           </form>
           <div className={styles.buttons}>
             <Button
               text={"저장"}
+              type={"submit"}
               form={"inputMachine"}
               disabled={products.length === 0 ? true : false}
             />
