--- src/pages/InputMachines.module.css
@@ -1,45 +1,0 @@
-.section {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.title {
-  position: sticky;
-  z-index: 999;
-  width: 100%;
-  top: 0;
-  text-align: center;
-  padding: 0.7rem 0;
-  font-size: xx-large;
-  font-weight: bold;
-  margin-bottom: 0.5rem;
-  background-color: var(--navbar-background-color);
-  color: var(--text-basic-color);
-}
-
-.empty {
-  display: flex;
-  justify-content: center;
-  margin-top: 12rem;
-  color: var(--navbar-background-color);
-  font-size: x-large;
-}
-
-.form {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  overflow: auto;
-  padding: 2rem 2.5rem;
-}
-
-.buttons {
-  display: flex;
-  position: sticky;
-  bottom: 0;
-  justify-content: space-around;
-  transform: translateY(2%);
-  margin-top: 1rem;
-  font-weight: bold;
-}
