+++ src/pages/RandomFoodTemp.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from "react";
-import styles from "./CustomMachineTemp.module.css";
+import styles from "./RandomTemp.module.css";
 import SetTemp from "../components/SetTemp";
-import { useCheatMachines, useCheatProducts } from "../hooks/useCheat";
+import { useCheatFoods, useCheatProducts } from "../hooks/useCheat";
 import ManagerList from "../components/ManagerList";
 import Banner from "../components/Banner";
 import Modal from "../components/Modal";
 
-export default function CustomMachineTemp() {
-  const { submitCustomTemp, data, setCustomTemp } = useCheatMachines();
+export default function RandomFoodTemp() {
+  const { submitCustomTemp, data, setCustomTemp } = useCheatFoods();
   const {
     handleSave,
     handleSubmit,
@@ -25,13 +25,13 @@ export default function CustomMachineTemp() {
   });
 
   useEffect(() => {
-    setProducts(data?.customCheatMachine);
+    setProducts(data?.customCheatFood);
   }, [data]);
 
   return (
     <section className={styles.section}>
       <div className={styles.title}>
-        <div className={styles.text}>기기 범위</div>
+        <div className={styles.text}>식품 범위</div>
         {data?.mgrList && (
           <ManagerList
             className={styles.mgrList}
@@ -98,7 +98,7 @@ export default function CustomMachineTemp() {
           }}
           disabled={success || warning}
         >
-          오전 기기 제출
+          오전 식품 제출
         </button>
         <button
           className={styles.submitBtn}
@@ -108,7 +108,7 @@ export default function CustomMachineTemp() {
           }}
           disabled={success || warning}
         >
-          오후 기기 제출
+          오후 식품 제출
         </button>
       </section>
     </section>
